pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fastapi-ulid-postgres"
version = "0.1.0"
description = "ULID demo"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.15.1",
    "asyncpg>=0.30.0",
    "fastapi[standard]>=0.115.11",
    "loguru>=0.7.3",
    "passlib>=1.7.4",
    "pydantic-settings>=2.8.1",
    "python-jose>=3.4.0",
    "python-ulid>=3.0.0",
    "sqlalchemy-utils>=0.41.2",
    "sqlalchemy[asyncio]>=2.0.39",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-env>=1.1.5",
    "pytest-cov>=6.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "isort>=5.13.0",
    "mypy>=1.8.0",
    "ruff>=0.2.0",
]

[tool.black]
line-length = 88

[tool.isort]
profile = "black"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
filterwarnings = [
    "error",
    "ignore::UserWarning",
    "ignore::DeprecationWarning",
    "ignore:.*unclosed.*:ResourceWarning",
]
env = [
    "API_ENVIRONMENT=pytest",
    "POSTGRES_DB=test",
]
asyncio_mode = "strict"
asyncio_default_fixture_loop_scope = "function"



[tool.hatch.build.targets.wheel]
packages = ["app/"]
